Committed : #2896

Related work items: #2896

+++ ERP-API/Domain/Entities/Purchase/PurchaseReturn.cs
@@ -348,7 +348,7 @@ namespace ERP_API.Domain.Entities.Purchase
 
         public string UnitName { get; set; }
 
-        public long? ItemUomBuyId { get; set; }
+        public int? ItemUomBuyId { get; set; }
 
         public string ItemUomBuyName { get; set; }
